requirements.txt
prettytable==3.10.0
scapy==2.5.0
wcwidth==0.2.13

python-dotenv~=1.0.1